+++ CHANGES.md
@@ -4,6 +4,7 @@
   * Cancellation of uploads
   * Added Request#withFile(String, byte[]), Request#withFile(String, ByteBuffer)
   * Added Request#withEntity(HttpEntity) and Request#withContent(String, String)
+  * Added PutResource example
 
 ## 1.0.0 2011-05-19